requirements.txt
flask
flask_socketio
biopython
numpy
scipy
